# [ci] Add benchmarks for PRs