--- a/.travis/benchmark.sh
+++ b/.travis/benchmark.sh
@@ -11,12 +11,14 @@ fi
 go get golang.org/x/perf/cmd/benchstat
 
 # Run benchmark against current branch
-echo "Running benchmarks on current commit..."
-go test -run NoTests -bench . -count 5 > /tmp/new
+echo "Running benchmarks on current branch PR ($TRAVIS_PULL_REQUEST_BRANCH)..."
+time go test -run NoTests -bench . -count 5 > /tmp/new
 
 # Run bebchnark against master
-git reset --hard $TRAVIS_BRANCH
 echo "Running benchmarks on $TRAVIS_BRANCH branch..."
-go test -run NoTests -bench . -count 5 > /tmp/master
+git reset --hard $TRAVIS_BRANCH
+time go test -run NoTests -bench . -count 5 > /tmp/master
 
+echo "#########################################################"
+echo "Results:"
 benchstat /tmp/master /tmp/new