--- a/.travis/benchmark.sh
+++ b/.travis/benchmark.sh
@@ -10,14 +10,15 @@ fi
 # Get utilities
 go get golang.org/x/perf/cmd/benchstat
 
-# Run benchmark against current branch
-echo "Running benchmarks on current branch PR ($TRAVIS_PULL_REQUEST_BRANCH)..."
-time go test -run NoTests -bench . -count 5 > /tmp/new
-
 # Run bebchnark against master
 echo "Running benchmarks on $TRAVIS_BRANCH branch..."
 git reset --hard $TRAVIS_BRANCH
 time go test -run NoTests -bench . -count 5 > /tmp/master
+
+# Run benchmark against current branch
+echo "Running benchmarks on current branch PR ($TRAVIS_PULL_REQUEST_BRANCH)..."
+git reset --hard $TRAVIS_PULL_REQUEST_BRANCH
+time go test -run NoTests -bench . -count 5 > /tmp/new
 
 echo "#########################################################"
 echo "Results:"