--- a/.travis/benchmark.sh
+++ b/.travis/benchmark.sh
@@ -11,15 +11,14 @@ fi
 # Get utilities
 go get golang.org/x/perf/cmd/benchstat
 
+# Run benchmark against current branch
+echo "Running benchmarks on PR branch $(git rev-parse HEAD)..."
+time go test -run NoTests -bench . -count 5 > /tmp/new
+
 # Run bebchnark against master
 echo "Running benchmarks on $TRAVIS_BRANCH branch..."
 git reset --hard $TRAVIS_BRANCH
 time go test -run NoTests -bench . -count 5 > /tmp/master
-
-# Run benchmark against current branch
-echo "Running benchmarks on PR branch ($TRAVIS_PULL_REQUEST_BRANCH)..."
-git checkout $TRAVIS_PULL_REQUEST_BRANCH
-time go test -run NoTests -bench . -count 5 > /tmp/new
 
 echo "#########################################################"
 echo "Results:"